Pri upisu rezultata user-u se upisuje statistika
Kada korisnik zavrsi sa testom, taj rezultat se dodaje u njegovu statistiku, tj. povecava se broj testova koje je uradio, i suma poena, odakle moze da se izvuce i prosecan broj poena na testovima

+++ App/FrontendKengur/app/src/main/java/com/example/kengur/dtos/response/LeaderboardResponse.kt
@@ -12,7 +12,7 @@ data class LeaderboardResponse(
     @SerializedName("user")
     var userDTO:UserDTO,
     @SerializedName("points")
-    var points:Float,
+    var points:Double,
     @SerializedName("class")
     var testClass:String
 )
